renaming "topology master" to "topology manager" in heron/healthmgr
